main Split platform-specific scanning code
